src/main/java/roboprotocolo/function/ValidaProtocoloFunction.java
package roboprotocolo.function;

import com.amazonaws.services.lambda.runtime.Context;
import com.amazonaws.services.lambda.runtime.RequestHandler;
import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;

public class ValidaProtocoloFunction implements RequestHandler <APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
    public APIGatewayProxyResponseEvent handleRequest(final APIGatewayProxyRequestEvent input, final Context context) {
        return new APIGatewayProxyResponseEvent().withStatusCode(200).withBody("OK");
    }
}
